Removed hello world controller example, and added method to check if a user is currently logged in
