+++ src/main/resources/static/index.html
@@ -16,5 +16,6 @@
     <button> Logout </button>
 </a>
 
+
 </body>
 </html>
